Use melib IMAP connection with own backend
This is one dirty mega commit that I will, unfortunately, not split into
individual commits. I've been working this for a long time that I
should get it over with.

The main goal of this commit is to add imap.rs, the IMAP backend
implementation which side-steps the state-keeping in the melib IMAP
backend.

All other changes attempt to take this backend implementation and make
some basic sync work.