Snapshot of Vala state
Since the Vala work is not being continued, this is not very important,
but I want to capture it in the git history in case I needed for further
reference. The work on this changeset is basically what pushed me to
rewrite everything in Rust. Libgda is pretty terrible and the API that
is currently shipped with distributions is pretty out of date.